+++ plugin/evm/config/config.go
@@ -9,6 +9,7 @@ import (
 	"time"
 
 	"github.com/ava-labs/avalanchego/utils/constants"
+	"github.com/ava-labs/coreth/core"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/spf13/cast"
@@ -47,7 +48,6 @@ const (
 	defaultLogJSONFormat                          = false
 	defaultMaxOutboundActiveRequests              = 16
 	defaultPopulateMissingTriesParallelism        = 1024
-	defaultRecentBlocksWindow                     = 1024
 	defaultStateSyncServerTrieCache               = 64 // MB
 	defaultAcceptedCacheSize                      = 32 // blocks
 
@@ -124,8 +124,11 @@ type Config struct {
 	PopulateMissingTries            *uint64 `json:"populate-missing-tries,omitempty"`   // Sets the starting point for re-populating missing tries. Disables re-generation if nil.
 	PopulateMissingTriesParallelism int     `json:"populate-missing-tries-parallelism"` // Number of concurrent readers to use when re-populating missing tries on startup.
 	PruneWarpDB                     bool    `json:"prune-warp-db-enabled"`              // Determines if the warpDB should be cleared on startup
-	HistoricalProofs                bool    `json:"historical-proofs"`                  // HistoricalProofs, if set to true, allows to query historical block numbers for proofs.
-	RecentBlocksWindow              uint64  `json:"recent-blocks-window"`               // RecentBlocksWindow is the number of blocks before the last accepted block to be considered as recent and non-historical. It defaults to 1024.
+
+	// HistoricalStateQueryWindow is the number of blocks before the last accepted block to be accepted for state queries.
+	// For archive nodes, it defaults to 43200 and can be set to 0 to indicate to accept any block query.
+	// For non-archive nodes, it is forcibly set to the value of [core.TipBufferSize].
+	HistoricalStateQueryWindow *uint64 `json:"historical-state-query-window,omitempty"`
 
 	// Metric Settings
 	MetricsExpensiveEnabled bool `json:"metrics-expensive-enabled"` // Debug-level metrics that might impact runtime performance
@@ -285,13 +288,38 @@ func (c *Config) SetDefaults(txPoolConfig TxPoolConfig) {
 	c.LogJSONFormat = defaultLogJSONFormat
 	c.MaxOutboundActiveRequests = defaultMaxOutboundActiveRequests
 	c.PopulateMissingTriesParallelism = defaultPopulateMissingTriesParallelism
-	c.RecentBlocksWindow = defaultRecentBlocksWindow
 	c.StateSyncServerTrieCache = defaultStateSyncServerTrieCache
 	c.StateSyncCommitInterval = defaultSyncableCommitInterval
 	c.StateSyncMinBlocks = defaultStateSyncMinBlocks
 	c.StateSyncRequestSize = DefaultStateSyncRequestSize
 	c.AllowUnprotectedTxHashes = defaultAllowUnprotectedTxHashes
 	c.AcceptedCacheSize = defaultAcceptedCacheSize
+}
+
+// UpdateWithDefaults updates the configuration with defaults depending
+// on fields already set.
+func (c *Config) UpdateWithDefaults() {
+	if c.Pruning {
+		// Force state query window to be the value of [core.TipBufferSize].
+		c.HistoricalStateQueryWindow = ptrTo(uint64(core.TipBufferSize))
+	} else {
+		// Default the state query window to approximately 24 hours of block history
+		// for nodes operating in archive mode.
+		const estimatedBlockAcceptPeriod = 2 * time.Second
+		const defaultBlocksWindow = uint64(24 * time.Hour / estimatedBlockAcceptPeriod)
+		c.HistoricalStateQueryWindow = defaultPtr(c.HistoricalStateQueryWindow, defaultBlocksWindow)
+	}
+}
+
+func ptrTo[T any](x T) *T { return &x }
+
+func defaultPtr[T any](existing *T, defaultValue T) (defaulted *T) {
+	if existing != nil {
+		return existing
+	}
+	defaulted = new(T)
+	*defaulted = defaultValue
+	return defaulted
 }
 
 func (d *Duration) UnmarshalJSON(data []byte) (err error) {