+++ plugin/evm/vm_test.go
@@ -405,6 +405,7 @@ func TestVMConfigDefaults(t *testing.T) {
 	vmConfig.SetDefaults(defaultTxPoolConfig)
 	vmConfig.RPCTxFeeCap = txFeeCap
 	vmConfig.EnabledEthAPIs = enabledEthAPIs
+	vmConfig.UpdateWithDefaults()
 	require.Equal(t, vmConfig, vm.config, "VM Config should match default with overrides")
 	require.NoError(t, vm.Shutdown(context.Background()))
 }
@@ -415,6 +416,7 @@ func TestVMNilConfig(t *testing.T) {
 	// VM Config should match defaults if no config is passed in
 	var vmConfig config.Config
 	vmConfig.SetDefaults(defaultTxPoolConfig)
+	vmConfig.UpdateWithDefaults()
 	require.Equal(t, vmConfig, vm.config, "VM Config should match default config")
 	require.NoError(t, vm.Shutdown(context.Background()))
 }